expand attachment workflows and refine execution progress

+++ Fleet.Server/Agents/PhaseResult.cs
@@ -20,7 +20,7 @@ public record PhaseResult(
     string? Error = null,
 
     /// <summary>Latest estimated completion percent reported during this attempt.</summary>
-    int EstimatedCompletionPercent = 0,
+    double EstimatedCompletionPercent = 0,
 
     /// <summary>Latest status summary reported during this attempt.</summary>
     string? LastProgressSummary = null
